Separate digittransit conf scoring to a new file
There's nothing common in scoring any more, so bring back the
upstram master and put our own code to a dedicated module.
New changes:
- Remove unused code from DT scoring module
- Drop the vague 'enrichment' scoring to make logic understandable
- Score address parts independently one by one with desired weights
- Skip plain name scoring only when search text is exactly street + number
- Don't check first item of parsed regions, because addressit puts the name part there
